+++ src/index.ts
@@ -1,3 +1,3 @@
-export { StaticKit, useStaticKit } from './context';
+export { StaticKitProvider, useStaticKit } from './context';
 export { useForm } from './useForm';
 export { ValidationError } from './ValidationError';